Combine overlay implementations to guarantee consistent size
